+++ src/main/java/com/ejoongseok/wmslive/outbound/domain/PackagingMaterialRepository.java
@@ -1,11 +1,11 @@
-package com.ejoongseok.wmslive.outbound.feature;
+package com.ejoongseok.wmslive.outbound.domain;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-class PackagingMaterialRepository {
+public class PackagingMaterialRepository {
     private final Map<Long, PackagingMaterial> packagingMaterialMap = new HashMap<>();
     private Long sequence = 1L;
